Split up mathematical.hpp and mathematical.cpp
