+++ tests/cpp/meson.build
@@ -9,11 +9,7 @@ tests_all = executable(
     [
         'nodes/indexing/test_advanced.cpp',
         'nodes/indexing/test_basic.cpp',
-        'nodes/mathematical/test_binaryop.cpp',
-        'nodes/mathematical/test_softmax.cpp',
-        'nodes/mathematical/test_naryop.cpp',
-        'nodes/mathematical/test_reduce.cpp',
-        'nodes/mathematical/test_unaryop.cpp',
+        'nodes/test_binaryop.cpp',
         'nodes/test_collections.cpp',
         'nodes/test_constants.cpp',
         'nodes/test_creation.cpp',
@@ -23,10 +19,14 @@ tests_all = executable(
         'nodes/test_lp.cpp',
         'nodes/test_lambda.cpp',
         'nodes/test_manipulation.cpp',
+        'nodes/test_naryop.cpp',
         'nodes/test_numbers.cpp',
         'nodes/test_quadratic_model.cpp',
+        'nodes/test_reduce.cpp',
+        'nodes/test_softmax.cpp',
         'nodes/test_sorting.cpp',
         'nodes/test_statistics.cpp',
+        'nodes/test_unaryop.cpp',
         'test_array.cpp',
         'test_functional.cpp',
         'test_graph.cpp',
